test(onboard): split dockerfile patch coverage
